notifications/requirements.txt
pytest==7.1.1
requests==2.31.0
psycopg2-binary

